renamed animation.py
