+++ projects/2023/project06_shallow_water/sw_gt4py/animation.py
